+++ lib/Traits/TableTimeStyleTraits.ts
@@ -45,4 +45,12 @@ export default class TableTimeStyleTraits extends ModelTraits {
     description: "Display duration for each row in the table, in minutes."
   })
   displayDuration?: number;
+
+  @primitiveTrait({
+    name: "Spread start time",
+    type: "boolean",
+    description:
+      'Indicates if start time of feature should be "spread" so that all features are displayed at the earliest time step. This is useful for non-contiguous sensor data. If true, the earliest time step will display the earliest values for all features (eg sensor IDs) - even if the time value is **after** the earliest time step.'
+  })
+  spreadStartTime?: boolean;
 }